service de ventas inicial, la estructura y lógica de detalles tanto en venta como compra se repiten

+++ src/main/java/com/togamma/heladeria/dto/response/compra/CompraResponseDTO.java
@@ -8,7 +8,7 @@ public record CompraResponseDTO(
     LocalDateTime createdAt,
     String descripcion,
     String numeroComprobante,
-    Long total,
+    Double total,
     String estado,
     String proveedor,
     List<DetCompraResponseDTO> detalles
